infra/airflow/requirements.txt
astronomer-cosmos==1.13.0
apache-airflow-providers-trino
apache-airflow-providers-docker
